requirements.txt
# Django framework
Django==5.2.5

# Django REST Framework
djangorestframework==3.19.0

# JWT authentication
djangorestframework-simplejwt==5.2.1

# PostgreSQL driver
psycopg2-binary==2.9.7

# Required by Django
asgiref==3.7.2
pytz==2025.7
sqlparse==0.5.5

# Optional: For handling images if you use ImageField
Pillow==10.1.0
